chrome/test/data/android/webvr_instrumentation/html/test_screen_taps_registered_on_cardboard.html

<!doctype html>
<!--
Tests that screen taps are registered while in VR when viewer is
Cardboard
-->
<html>
  <head>
    <link rel="stylesheet" type="text/css" href="../resources/webvr_e2e.css">
  </head>
  <body>
    <canvas id="webgl-canvas"></canvas>
    <script src="../../../../../../third_party/WebKit/LayoutTests/resources/testharness.js"></script>
    <script src="../resources/webvr_e2e.js"></script>
    <script src="../resources/webvr_boilerplate.js"></script>
    <script>
      var t = async_test("Screen taps registered when in VR w/ Cardboard");
      window.addEventListener("vrdisplaypresentchange",
          () => {finishJavascriptStep();}, false);
      var numTaps = 0;
      webglCanvas.addEventListener("click", () => {numTaps++;}, false);

      function stepVerifyNoInitialTaps() {
        t.step( () => {
          assert_equals(numTaps, 0, "No initial taps");
        });
        finishJavascriptStep();
      }

      function stepVerifyAdditionalTap() {
          t.step_func_done( () => {
            // We expect 1 tap from entering VR + 1 additional one after
            assert_equals(numTaps, 2,
                "Two taps registered after two taps given");
          })();
      }

    </script>
  </body>
</html>
